+++ clients/js/packages/starshipjs/src/hooks.ts
@@ -3,7 +3,7 @@ import fs from 'fs';
 import yaml from 'js-yaml';
 import fetch from 'node-fetch';
 
-import { ChainConfig, ConfigContext } from './config';
+import { type ChainConfig, ConfigContext } from './config';
 
 export const useRegistry = async (configFile: string): Promise<ChainRegistryFetcher> => {
   const config = yaml.load(fs.readFileSync(configFile, 'utf8')) as ChainConfig;
@@ -12,8 +12,8 @@ export const useRegistry = async (configFile: string): Promise<ChainRegistryFetc
   const urls: string[] = [];
   config.chains?.forEach((chain) => {
     urls.push(
-      `${registryUrl}/chains/${chain.name}`,
-      `${registryUrl}/chains/${chain.name}/assets`
+      `${registryUrl}/chains/${chain.id}`,
+      `${registryUrl}/chains/${chain.id}/assets`
     );
   });
   config.relayers?.forEach((relayer) => {
@@ -41,14 +41,14 @@ export const useChain = (chainName: string) => {
   const chainInfo = registry!.getChainInfo(chainName);
   const chainID = chainInfo.chain.chain_id;
 
-  const getRpcEndpoint = () => {
+  const getRpcEndpoint = async () => {
     return `http://localhost:${
-      config.chains.find((chain) => chain.name === chainID)!.ports.rpc
+      config.chains.find((chain) => chain.id === chainID)!.ports.rpc
     }`;
   };
-  const getRestEndpoint = () => {
+  const getRestEndpoint = async () => {
     return `http://localhost:${
-      config.chains.find((chain) => chain.name === chainID)!.ports.rest
+      config.chains.find((chain) => chain.id === chainID)!.ports.rest
     }`;
   };
 
@@ -59,16 +59,16 @@ export const useChain = (chainName: string) => {
     return data['genesis'][0]['mnemonic'];
   };
 
-  const getCoin = () => {
+  const getCoin = async () => {
     return chainInfo.fetcher.getChainAssetList(chainName).assets[0];
   };
 
   const creditFromFaucet = async (address: string, denom: string | null = null) => {
     const faucetEndpoint = `http://localhost:${
-      config.chains.find((chain) => chain.name === chainID)!.ports.faucet
+      config.chains.find((chain) => chain.id === chainID)!.ports.faucet
     }/credit`;
     if (!denom) {
-      denom = getCoin().base;
+      denom = (await getCoin()).base;
     }
     await fetch(faucetEndpoint, {
       method: 'POST',