+++ clients/js/packages/client/CHANGELOG.md
@@ -3,6 +3,66 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.7.0](https://github.com/cosmology-tech/starship/compare/@starship-ci/client@2.6.1...@starship-ci/client@2.7.0) (2024-06-19)
+
+**Note:** Version bump only for package @starship-ci/client
+
+## [2.6.1](https://github.com/cosmology-tech/starship/compare/@starship-ci/client@2.6.0...@starship-ci/client@2.6.1) (2024-06-14)
+
+**Note:** Version bump only for package @starship-ci/client
+
+# [2.6.0](https://github.com/cosmology-tech/starship/compare/@starship-ci/client@2.5.5...@starship-ci/client@2.6.0) (2024-06-13)
+
+**Note:** Version bump only for package @starship-ci/client
+
+## [2.5.5](https://github.com/cosmology-tech/starship/compare/@starship-ci/client@2.5.4...@starship-ci/client@2.5.5) (2024-06-13)
+
+**Note:** Version bump only for package @starship-ci/client
+
+## [2.5.4](https://github.com/cosmology-tech/starship/compare/@starship-ci/client@2.5.3...@starship-ci/client@2.5.4) (2024-06-13)
+
+**Note:** Version bump only for package @starship-ci/client
+
+## [2.5.3](https://github.com/cosmology-tech/starship/compare/@starship-ci/client@2.5.2...@starship-ci/client@2.5.3) (2024-06-13)
+
+**Note:** Version bump only for package @starship-ci/client
+
+## [2.5.2](https://github.com/cosmology-tech/starship/compare/@starship-ci/client@2.5.1...@starship-ci/client@2.5.2) (2024-06-13)
+
+**Note:** Version bump only for package @starship-ci/client
+
+## [2.5.1](https://github.com/cosmology-tech/starship/compare/@starship-ci/client@2.5.0...@starship-ci/client@2.5.1) (2024-06-13)
+
+**Note:** Version bump only for package @starship-ci/client
+
+# [2.5.0](https://github.com/cosmology-tech/starship/compare/@starship-ci/client@2.4.0...@starship-ci/client@2.5.0) (2024-06-13)
+
+**Note:** Version bump only for package @starship-ci/client
+
+# [2.4.0](https://github.com/cosmology-tech/starship/compare/@starship-ci/client@2.3.0...@starship-ci/client@2.4.0) (2024-06-12)
+
+**Note:** Version bump only for package @starship-ci/client
+
+# [2.3.0](https://github.com/cosmology-tech/starship/compare/@starship-ci/client@2.2.0...@starship-ci/client@2.3.0) (2024-06-12)
+
+**Note:** Version bump only for package @starship-ci/client
+
+# [2.2.0](https://github.com/cosmology-tech/starship/compare/@starship-ci/client@2.1.0...@starship-ci/client@2.2.0) (2024-06-10)
+
+**Note:** Version bump only for package @starship-ci/client
+
+# [2.1.0](https://github.com/cosmology-tech/starship/compare/@starship-ci/client@2.0.2...@starship-ci/client@2.1.0) (2024-06-07)
+
+**Note:** Version bump only for package @starship-ci/client
+
+## [2.0.2](https://github.com/cosmology-tech/starship/compare/@starship-ci/client@2.0.1...@starship-ci/client@2.0.2) (2024-06-07)
+
+**Note:** Version bump only for package @starship-ci/client
+
+## [2.0.1](https://github.com/cosmology-tech/starship/compare/@starship-ci/client@2.0.1...@starship-ci/client@2.0.1) (2024-06-07)
+
+**Note:** Version bump only for package @starship-ci/client
+
 ## [2.0.1](https://github.com/cosmology-tech/starship/compare/@starship-ci/client@1.6.0...@starship-ci/client@2.0.1) (2024-05-23)
 
 **Note:** Version bump only for package @starship-ci/client