+++ clients/js/packages/starshipjs/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.3.0](https://github.com/cosmology-tech/starship/compare/starshipjs@2.2.0...starshipjs@2.3.0) (2024-06-13)
+
+**Note:** Version bump only for package starshipjs
+
+# [2.2.0](https://github.com/cosmology-tech/starship/compare/starshipjs@2.1.0...starshipjs@2.2.0) (2024-06-10)
+
+**Note:** Version bump only for package starshipjs
+
+# [2.1.0](https://github.com/cosmology-tech/starship/compare/starshipjs@2.0.2...starshipjs@2.1.0) (2024-06-07)
+
+**Note:** Version bump only for package starshipjs
+
+## [2.0.2](https://github.com/cosmology-tech/starship/compare/starshipjs@2.0.1...starshipjs@2.0.2) (2024-06-07)
+
+**Note:** Version bump only for package starshipjs
+
+## [2.0.1](https://github.com/cosmology-tech/starship/compare/starshipjs@2.0.1...starshipjs@2.0.1) (2024-06-07)
+
+**Note:** Version bump only for package starshipjs
+
 ## [2.0.1](https://github.com/cosmology-tech/starship/compare/starshipjs@1.4.0...starshipjs@2.0.1) (2024-05-23)
 
 **Note:** Version bump only for package starshipjs